# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTaskRequest": {
        "properties": {
          "command": {
            "type": "string"
          },
          "enabled": {
            "type": [
              "boolean",
              "null"
            ]
          },
          "name": {
            "type": "string"
          },
          "schedule": {
            "type": [
              "string",
              "null"
            ]
          }
        },
        "required": [
          "name",
          "command"
        ],
        "type": "object"
      },
...
      "CreateWebsiteRequest": {
        "properties": {
          "website": {
            "$ref": "#/components/schemas/Website"
          }
        },
        "required": [
          "website"
        ],
        "type": "object"
      },
...
      "PaginatedResponse<ScheduledTask>_ScheduledTask": {
        "properties": {
          "data": {
            "items": {
              "properties": {
                "command": {
                  "type": "string"
                },
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "enabled": {
                  "type": "boolean"
                },
                "id": {
                  "format": "int64",
                  "type": "integer"
                },
                "last_output": {
                  "type": "string"
                },
                "last_run_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "last_status": {
                  "description": "never | success | failed",
                  "type": "string"
                },
                "name": {
                  "type": "string"
                },
                "next_run_at": {
                  "format": "date-time",
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "schedule": {
                  "description": "标准 5 字段 cron 表达式",
                  "type": "string"
                },
                "updated_at": {
                  "format": "date-time",
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "command",
                "schedule",
                "enabled",
                "last_status",
                "last_output",
                "created_at",
                "updated_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "type": "object"
      },
...
      "PaginatedResponse<Website>_Website": {
        "properties": {
          "data": {
            "items": {
              "properties": {
                "created_at": {
                  "format": "date-time",
                  "type": "string"
                },
                "domain": {
                  "type": "string"
                },
                "engine": {
                  "type": "string"
                },
                "id": {
                  "format": "int64",
                  "type": "integer"
                },
                "name": {
                  "type": "string"
                },
                "node_id": {
                  "format": "int64",
                  "type": "integer"
                },
                "proxy_enabled": {
                  "type": "boolean"
                },
                "proxy_pass": {
                  "type": [
                    "string",
                    "null"
                  ]
                },
                "resource_version": {
                  "description": "乐观并发控制版本号：写入时需匹配当前值，冲突返回稳定 `CONFLICT`。",
                  "format": "int64",
                  "type": "integer"
                },
                "root_path": {
                  "type": "string"
                },
                "ssl_enabled": {
                  "type": "boolean"
                },
                "status": {
                  "type": "string"
                }
              },
              "required": [
                "id",
                "name",
                "domain",
                "root_path",
                "status",
                "node_id",
                "engine",
                "ssl_enabled",
                "proxy_enabled",
                "created_at"
              ],
              "type": "object"
            },
            "type": "array"
          },
          "page": {
            "type": "integer"
          },
          "page_size": {
            "type": "integer"
          },
          "total": {
            "type": "integer"
          },
          "total_pages": {
            "type": "integer"
          }
        },
        "required": [
          "data",
          "page",
          "page_size",
          "total",
          "total_pages"
        ],
        "type": "object"
      },
...
      "ScheduledTask": {
        "properties": {
          "command": {
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "enabled": {
            "type": "boolean"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "last_output": {
            "type": "string"
          },
          "last_run_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "last_status": {
            "description": "never | success | failed",
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "next_run_at": {
            "format": "date-time",
            "type": [
              "string",
              "null"
            ]
          },
          "schedule": {
            "description": "标准 5 字段 cron 表达式",
            "type": "string"
          },
          "updated_at": {
            "format": "date-time",
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "command",
          "schedule",
          "enabled",
          "last_status",
          "last_output",
          "created_at",
          "updated_at"
        ],
        "type": "object"
      },
...
      "Website": {
        "properties": {
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "domain": {
            "type": "string"
          },
          "engine": {
            "type": "string"
          },
          "id": {
            "format": "int64",
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "node_id": {
            "format": "int64",
            "type": "integer"
          },
          "proxy_enabled": {
            "type": "boolean"
          },
          "proxy_pass": {
            "type": [
              "string",
              "null"
            ]
          },
          "resource_version": {
            "description": "乐观并发控制版本号：写入时需匹配当前值，冲突返回稳定 `CONFLICT`。",
            "format": "int64",
            "type": "integer"
          },
          "root_path": {
            "type": "string"
          },
          "ssl_enabled": {
            "type": "boolean"
          },
          "status": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "domain",
          "root_path",
          "status",
          "node_id",
          "engine",
          "ssl_enabled",
          "proxy_enabled",
          "created_at"
        ],
        "type": "object"
      }
    },
    "securitySchemes": {
      "BearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "FlamePanel 面板后端 REST API（OpenAPI 3.x，由 utoipa 编译期生成）",
    "license": {
      "name": ""
    },
    "title": "FlamePanel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/operation-logs/export": {
      "get": {
        "description": "Stage4.4：支持 CSV / JSON 两种格式，便于归档与 SIEM 对接。\nCSV 带 BOM（UTF-8）以兼容 Excel；JSON 为 UTF-8 数组。",
        "operationId": "export",
        "parameters": [
          {
            "in": "query",
            "name": "format",
            "required": false,
            "schema": {
              "type": [
                "string",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "导出文件"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "`GET /api/operation-logs/export?format=csv|json` — 导出全部审计日志",
        "tags": [
          "operation-logs"
        ]
      }
    },
...
    "/api/scheduled-tasks": {
      "get": {
        "operationId": "list_tasks",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int64",
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "每页条数（1~200）",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse<ScheduledTask>_ScheduledTask"
                }
              }
            },
            "description": "任务列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "定时任务列表（分页）",
        "tags": [
          "scheduled_tasks"
        ]
      },
      "post": {
        "operationId": "create_task",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduledTask"
                }
              }
            },
            "description": "创建成功"
          },
          "400": {
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "创建定时任务",
        "tags": [
          "scheduled_tasks"
        ]
      }
    },
...
    "/api/websites": {
      "get": {
        "operationId": "list_websites",
        "parameters": [
          {
            "description": "页码（从 1 开始）",
            "in": "query",
            "name": "page",
            "required": false,
            "schema": {
              "default": 1,
              "format": "int64",
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          },
          {
            "description": "每页条数（1~200）",
            "in": "query",
            "name": "page_size",
            "required": false,
            "schema": {
              "default": 20,
              "format": "int64",
              "maximum": 200,
              "minimum": 1,
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse<Website>_Website"
                }
              }
            },
            "description": "网站列表"
          },
          "401": {
            "description": "未认证"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "网站列表（分页）",
        "tags": [
          "websites"
        ]
      },
      "post": {
        "operationId": "create_website",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebsiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "text/plain": {
                "schema": {
                  "format": "int64",
                  "type": "integer"
                }
              }
            },
            "description": "创建成功，返回网站 ID"
          },
          "400": {
            "description": "参数错误"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "创建网站",
        "tags": [
          "websites"
        ]
      }
    },
    "/api/websites/{id}": {
      "delete": {
        "operationId": "delete_website",
        "parameters": [
          {
            "description": "网站 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "删除成功"
          },
          "404": {
            "description": "网站不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "删除网站",
        "tags": [
          "websites"
        ]
      },
      "get": {
        "operationId": "get_website",
        "parameters": [
          {
            "description": "网站 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Website"
                }
              }
            },
            "description": "网站详情"
          },
          "404": {
            "description": "网站不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "获取网站详情",
        "tags": [
          "websites"
        ]
      },
      "put": {
        "operationId": "update_website",
        "parameters": [
          {
            "description": "网站 ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "int64",
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWebsiteRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Website"
                }
              }
            },
            "description": "更新成功"
          },
          "404": {
            "description": "网站不存在"
          }
        },
        "security": [
          {
            "BearerAuth": []
          }
        ],
        "summary": "更新网站",
        "tags": [
          "websites"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "默认",
      "url": "/"
    }
  ]
}